opencore:Fix corrupted wav file crash opencore
Issue 1: When playing corrupted wav files, audiotrack creation fails and
audio thread created exits with an error. The error reporting function
is not implemented. Also the function that starts the thread blocks on
semaphone in a while 1 loop. The audiothread that exits due to an error
condition doesn't signal the semaphore before exit.

Issue 2:Read on a corrupted wav file causes an exception. This is not being
handled.

CRs-Fixed: 222061