Opencore: Fix for displaying first few frames of low resolution clips with correct size
Fix to pass video width, height, display width and height using
correct key from the video decoder node to the surface output MIO.

Application registers a callback to be notififed about the video size.
This callback is invoked from the MIO on successful creation of overlay.

Passing the video and display parameters to the surface output MIO using
incorrect key in Prepare stage of video decoder node delays creation of overlay
and thus delays the notification to the application.

Change-Id: I601b0679b7a8f60ec3e6466d511b760311c9d96c
CRs-Fixed: 222406