Opencore: Ignoring frame duration if timestamps are not in order
Duration was not being calculated correctly in the parser in cases where
the clip contained B or P frames. The timestamps in these cases were not
in order, and the miscalculated duration would cause issues where frames
would drop. This change will make the parser ignore the duration by
setting it to 0 only if the timestamps are not in order.

Change-Id: I9177d02be1f931e5e0a367a947b23adbfaaecade
CRs-Fixed: 228020